Add information and explanations
1. information and explanations on input fields, info button with reference to open-
source code and simulation button to start the simulation
To improve user-friendliness, explanatory information is added to the input fields.
information is added to the input fields. In addition, the info button will link to the corresponding section
of the open source code. A “Start simulation” button allows users to
users to start the simulation only after the data has been entered completely.
2 “Other” buildings including heating
For the “Other buildings” category, the input screen will be expanded to include the option of
heating loads and flow temperatures.